feat: 서버에서 세션 가져오도록 수정 및 코드 리팩토링
