+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import AuthProvider from "@/lib/providers/AuthProvider";
 import "@/styles/globals.css";
 import "@/styles/reset.css";
 import type { Metadata } from "next";
+import { getServerSession } from "next-auth";
 import { ThemeProvider } from "next-themes";
 import { MoneygraphyRounded } from "./fonts";
 
@@ -10,17 +11,19 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const session = await getServerSession();
+
   return (
     <html lang="ko" suppressHydrationWarning>
       <body
-        className={`${MoneygraphyRounded.variable} antialiased text-dark dark:text-light`}
+        className={`${MoneygraphyRounded.className} antialiased text-dark dark:text-light`}
       >
-        <AuthProvider>
+        <AuthProvider session={session}>
           <ThemeProvider>{children}</ThemeProvider>
         </AuthProvider>
       </body>
