+++ src/lib/providers/AuthProvider.tsx
@@ -1,14 +1,15 @@
 "use client";
 
-import React from "react";
+import type { Session } from "next-auth";
 import { SessionProvider } from "next-auth/react";
 
 interface Props {
+  session: Session | null;
   children: React.ReactNode;
 }
 
-const AuthProvider = ({ children }: Props) => {
-  return <SessionProvider>{children}</SessionProvider>;
+const AuthProvider = ({ session, children }: Props) => {
+  return <SessionProvider session={session}>{children}</SessionProvider>;
 };
 
 export default AuthProvider;
